Renaming RouterModule.ts to lower case
